Changed type of PatternCountType in states_patterns.hpp from float to double - Gelfand-Ghosh stats can be way off if float values are used to keep track of pattern counts. Specifically, the total count ends up not being correct (i.e. not equal to the sum of the individual counts) for the compromise action and mean posterior predictive data sets.
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@565 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/states_patterns.hpp
@@ -25,7 +25,7 @@
 typedef std::vector<int8_t>				VecStateList;
 typedef std::vector<unsigned>			VecStateListPos;
 
-typedef float										PatternCountType;
+typedef double										PatternCountType;
 typedef	std::map<VecStateList, PatternCountType>	PatternMapType;
 typedef	std::vector<PatternCountType>				CountVectorType;
